+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.7
+- fix update, BER encode and decode
+## 0.0.5
+- all oid can be request, and respond PDU, can get value PDU.value;
 ## 0.0.4
 - Initial version, created by AlmPazel